3.5.0
Moved the projects collection to the tile view & the applications
collection to the application view (instead of the main view and storing
copies in the other two). This change was made with a view to
simplifying the code structure and hopefully make coding the admin
module easier.

This is also the commit before attempting to implement XML
serialisation.