3.5.1
XML serialisation implemented which has removed a lot of xml
loading/saving code.
Changed many of the messages classes and what messages are used to try
to simplify the whole process.

Having issues with icons (paths) displaying on anything but the first
tab of a tabcontrol. Seems like a bug. Have replicated the issue in a
simple app. Posted on Stackoverflow but not hopefully of a resolution.

Commit before removing the tabcontrol in the hope it fixes the icon
issue. Don't think the tabcontrol is required anyway.