3.6.1
Projects can now be deleted via the Edit view
The Edit view Cancel button now works
Settings file is delivered blank (apart from Application details). A
backup of the file is made when the app closes. This backup is loaded if
a new blank version is installed (when the app is updated). There is
also an option to load a settings file manually.
Commit when formally publishing to network.