3.8.1.0
Moved most of the Styles & code-behind from the App.xaml into the
relevant View in an attempt to reduce the amount of messages being send.
Didn't actually remove many but does make the code a little easier to
understand. Many of the Styles were only used in one View and therefore
didn't need to be in the App.xaml.

A few other bug fixes. Still have an issue where if a project has
multiple applications associated it launches all of them (not just the
selected application)